pyproject.toml
[tool.poetry]
name = "stl-dsa"
version = "0.1.0"
description = "A Django + Wagtail based website for the St Louis Chapter of DSA"
authors = ["Tyler Schlichenmeyer <tyler.schlichenmeyer@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2"
django-allauth = "^0.45.0"
django-extensions = "^2.2.9"
django-crispy-forms = "^1.9.1"
django-environ = "^0.4.5"
whitenoise = "^5.1.0"
django-rest-framework = "^0.1.0"
wagtail = "^2.14"
django-phonenumber-field = { extras = ["phonenumbers"], version = "^4.0.0" }
argon2-cffi = "^20.1.0"
wagtailmenus = "^3.0.2"
django-anymail = "^8.1"
gunicorn = "^20.0.4"
django-redis = "^4.12.1"
django-scheduler = "^0.9.3"
wagtail-blocks = "^0.7.2"
wagtailfontawesome = "^1.2.1"
requests = "^2.26.0"
python-dateutil = "^2.8.2"
django-configurations = "^2.2"
dj-database-url = "^0.5.0"
django-cache-url = "^3.2.3"
wagtail-localize = {version = "1.0rc3", allow-prereleases = true}
notebook = "^6.4.3"

[tool.poetry.dev-dependencies]
psycopg2-binary = "^2.8.5"
pytest-django = "^4.4.0"
factory-boy = "^3.2.0"
pytest-vcr = "^1.0.2"
black = "^20.8b1"
django-seed = "^0.2.2"
Faker = "^8.10.1"
pytest-selenium = "^2.0.1"
pytest = "^6.2.4"
flake8 = "^3.9.2"
django-debug-toolbar = "^3.2.2"
pytest-drf = "^1.1.2"
pre-commit = "^2.14.0"
pytest-cov = "^2.12.1"
django-coverage-plugin = "^2.0.0"
vcrpy = "^4.1.1"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
DJANGO_SETTINGS_MODULE = "config.settings"
DJANGO_CONFIGURATION = "Test"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
filterwarnings = [
    "ignore::django.utils.deprecation.RemovedInDjango40Warning",
    "ignore::django.utils.deprecation.RemovedInDjango41Warning"
]

[tool.flake8]
max-line-length = 88
select = "C,E,F,W,B,B950"
extend-ignore = "E203, E501"

[tool.black]
exclude = '^.*\b(migrations)\b.*$'

[build-system]
requires = ["poetry>=1.1.6"]
build-backend = "poetry.masonry.api"
